cookies and JET tokens implemention in login and logout
