+++ client/homePage.html
@@ -8,7 +8,7 @@
     <link href="https://fonts.googleapis.com/css?family=Comfortaa|Pacifico|Righteous|Teko" rel="stylesheet">
     <link rel="stylesheet"
           href="https://fonts.googleapis.com/css2?family=Material+Symbols+Outlined:opsz,wght,FILL,GRAD@20..48,100..700,0..1,-50..200"/>
-    <link href="../client/style.css" rel="stylesheet">
+    <link href="../clientPublic/style.css" rel="stylesheet">
     <title>Home Page</title>
 </head>
 <body>
@@ -134,7 +134,6 @@
         </ul>
     </div>
 </section>
-<script src="../client/scriptsHome.js"></script>
-<script src="../client/script.js"></script>
+<script src="../clientPublic/script.js"></script>
 </body>
 </html>
