+++ controllers/login_controller.js
@@ -16,30 +16,13 @@ const loginControl = async (req, res, next)=> {
             await loginService.handleLogin(req, res, next);
             const userFind = await dbHandler.getUserByEmail(req.body.email);
             const user = new userClass(userFind._id, userFind.type);
-            console.log(user);
             const accessToken = jwt.sign({user}, process.env.ACCESS_TOKEN_SECRET, { expiresIn: '15m' })
             res.cookie('token', accessToken, {httponly:true});
             res.status(200).json({});
     } catch (err) {
         console.log(err);
-       // return res.status(401).json({message: err.message});
+        return res.status(401).json({message: err.message});
     }
-    /*next();*/
-}
-
-const authenticateToken = (req,res,next)=>{
-    console.log("auth");
-/*    if (req.cookie.token===null) {
-        console.log(req.cookie.token);
-        return res.sendStatus(401);
-    }
-    jwt.verify(token,process.env.ACCESS_TOKEN_SECRET,(err,user)=>{
-        if (err){
-            console.log(err);
-            return res.sendStatus(403);  //token is no longer valid
-        }
-        next();
-    })*/
 }
 
 
@@ -57,4 +40,4 @@ const Permissions = async (req, res, next) => {
 }
 
 
-module.exports={Permissions,loginControl,authenticateToken};
+module.exports={Permissions,loginControl};
