+++ controllers/logout_controller.js
@@ -1,5 +1,4 @@
 exports.logout = (req, res) => {
-    console.log("HERE");
     res.cookie("token", "", { maxAge: 1 });
     res.end();
 };
