+++ data/dbHandler.js
@@ -24,7 +24,6 @@ async function addDoc(obj) {
 function typeUser(user) {
     let domain = user.email.split("@");
     domain = domain[1].split(".");
-    console.log(domain)
     if (domain.find(element => element === "shenkar")) {
         return "admin"
     } else {
@@ -35,7 +34,6 @@ function typeUser(user) {
  const addUser= async (user) => {
     if (user.password){
         user.password = await bcrypt.hash(user.password, 12);}
-     console.log(user.email)
     const domain = typeUser(user)
     const newUser = new User({
         "name": user.name,
